created reusable components/formatted components to reduce component size
