+++ client/components/Admin-Home.js
@@ -1,9 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
-import AllProducts from './index'
 import {BrowserRouter as Router, Link, Route, Switch} from 'react-router-dom'
-import {fetchAllOrders, fetchSingleOrders} from '../store/orders'
 import {fetchAllUsers} from '../store/user'
 
 /**
@@ -15,71 +13,27 @@ export class AdminHome extends React.Component {
   }
 
   componentDidMount() {
-    this.props.fetchAllOrders()
     this.props.fetchAllUsers()
   }
 
   render() {
-    console.log(this.props)
     const user = this.props.user.user
-    const orderArr = this.props.order.orders
-    const thisUsersOrder = orderArr.filter(order => {
-      return order.userId === user.id
-    })
-    const userArr = this.props.user.allUsers
-    console.log(userArr)
-    // console.log(thisUsersOrder[0])
+
     return (
       <div>
         <div>
           <h1>Welcome, {`${user.firstName} ${user.lastName}!`}</h1>
           <h4>{user.email}</h4>
         </div>
-        <div>
-          <h2>Orders</h2>
-          {thisUsersOrder && thisUsersOrder.length > 0 ? (
-            thisUsersOrder.map(order => {
-              return (
-                <div key={order.id}>
-                  <Link to={`/home/orders/${order.id}`}>
-                    <h4>Order Number: {order.id}</h4>
-                  </Link>
-                  <h4>Payment Type: {order.payment}</h4>
-                  <h4>Order Total:{order.total}</h4>
-                </div>
-              )
-            })
-          ) : (
-            <div>
-              <h4>...No Orders Placed Yet...</h4>
-            </div>
-          )}
-        </div>
-        <div>
-          <h2>Manage Users</h2>
 
-          {userArr && userArr.length > 0 ? (
-            userArr.map(user => {
-              return (
-                <div key={user.id}>
-                  <p>User Id: {user.id}</p>
-                  <p>User First Name: {user.firstName}</p>
-                  <p>User Last Name: {user.lastName}</p>
-                  <p>User Email: {user.email}</p>
-                  <p>Admin Status: {user.admin.toString()}</p>
-                  <label>
-                    <small>Toggle Admin Status:</small>
-                  </label>
-                  <select>
-                    <option>True</option>
-                    <option>False</option>
-                  </select>
-                </div>
-              )
-            })
-          ) : (
-            <div />
-          )}
+        <div>
+          <h2>Admin Portal Options</h2>
+          <Link to="/home/manageusers">
+            <button>Manage Users</button>
+          </Link>
+          <Link to="/home/orders">
+            <button>Order History</button>
+          </Link>
         </div>
       </div>
     )
@@ -91,15 +45,12 @@ export class AdminHome extends React.Component {
  */
 const mapState = state => {
   return {
-    user: state.user,
-    order: state.ordersReducer
+    user: state.user
   }
 }
 
 const mapDispatch = dispatch => {
   return {
-    fetchAllOrders: () => dispatch(fetchAllOrders()),
-    fetchSingleOrders: id => dispatch(fetchSingleOrders(id)),
     fetchAllUsers: () => dispatch(fetchAllUsers())
   }
 }
