+++ client/routes.js
@@ -9,7 +9,8 @@ import {Signup} from './components/SignUp'
 import {Cart} from './components/Cart'
 import OrderDetails from './components/OrderDetails'
 import AdminHome from './components/Admin-Home'
-
+import AdminManageUsers from './components/Admin_ManageUsers'
+import OrderHistory from './components/OrderHistory'
 /**
  * COMPONENT
  */
@@ -21,8 +22,8 @@ class Routes extends Component {
   render() {
     const {isLoggedIn} = this.props
     const {isAdmin} = this.props
-    console.log('isadmin', this.props.isAdmin)
-    console.log('islogged', this.props.isLoggedIn)
+    // console.log('isadmin', this.props.isAdmin)
+    // console.log('islogged', this.props.isLoggedIn)
     return (
       <Switch>
         {/* Routes placed here are available to all visitors */}
@@ -33,8 +34,18 @@ class Routes extends Component {
         <Route path="/cart" component={Cart} />
         {isAdmin && (
           <Switch>
-            <Route path="/home/admin" component={AdminHome} />
-            <Route path="/home/orders/:orderId" component={OrderDetails} />
+            <Route exact path="/home/admin" component={AdminHome} />
+            <Route
+              exact
+              path="/home/orders/:orderId"
+              component={OrderDetails}
+            />
+            <Route
+              exact
+              path="/home/manageusers"
+              component={AdminManageUsers}
+            />
+            <Route exact path="/home/orders" component={OrderHistory} />
           </Switch>
         )}
         {isLoggedIn && (
@@ -46,6 +57,7 @@ class Routes extends Component {
               path="/home/orders/:orderId"
               component={OrderDetails}
             />
+            <Route exact path="/home/orders" component={OrderHistory} />
           </Switch>
         )}
